recipes/requirements.txt
streamlit
mysql-connector-python
streamlit-option-menu  # If using for navigation
requests
Pillow
